push

+++ backend/api/views.py
@@ -19,6 +19,7 @@ class TagViewSet(viewsets.ReadOnlyModelViewSet):
     Получение списка тегов /
     конкретного тега
     """
+
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
     permission_classes = [IsAdminOrReadOnly]
@@ -29,11 +30,12 @@ class IngredientsViewSet(viewsets.ReadOnlyModelViewSet):
     Получение списка ингредиентов /
     конкретного ингредиента
     """
+
     queryset = Ingredient.objects.all()
     serializer_class = IngredientSerializer
     permission_classes = [IsAdminOrReadOnly]
     filter_backends = [IngredientSearchFilter]
-    search_fields = ['^name']
+    search_fields = ["^name"]
 
 
 class RecipeViewSet(viewsets.ModelViewSet):
@@ -41,6 +43,7 @@ class RecipeViewSet(viewsets.ModelViewSet):
     Получение списка рецептов /
     конкретного рецепта /
     """
+
     serializer_class = RecipeSerializer
     queryset = Recipe.objects.all()
     pagination_class = CustomPagination
@@ -50,41 +53,30 @@ class RecipeViewSet(viewsets.ModelViewSet):
     def perform_create(self, serializer):
         serializer.save(author=self.request.user)
 
-    @action(
-        methods=['GET'],
-        detail=False,
-        permission_classes=[IsAuthenticated]
-    )
+    @action(methods=["GET"], detail=False, permission_classes=[IsAuthenticated])
     def download_shopping_cart(self, request):
         """Метод для скачивания списка покупок"""
         shopping_result = {}
         ingredients = IngredientRecipes.objects.filter(
             recipes__shoppinglist__user=request.user
-        ).values_list(
-            'ingredients__name',
-            'ingredients__measurement_unit',
-            'amount'
-        )
+        ).values_list("ingredients__name", "ingredients__measurement_unit", "amount")
 
         for ingredient in ingredients:
             name = ingredient[0]
             if name not in shopping_result:
                 shopping_result[name] = {
-                    'measurement_unit': ingredient[1],
-                    'amount': ingredient[2]
+                    "measurement_unit": ingredient[1],
+                    "amount": ingredient[2],
                 }
             else:
-                shopping_result[name]['amount'] += ingredient[2]
+                shopping_result[name]["amount"] += ingredient[2]
 
         shopping_itog = (
-            f"{name} - {value['amount']} "
-            f"{value['measurement_unit']}\n"
+            f"{name} - {value['amount']} " f"{value['measurement_unit']}\n"
             for name, value in shopping_result.items()
         )
-        response = HttpResponse(shopping_itog, content_type='text/plain')
-        response[
-            'Content-Disposition'
-        ] = 'attachment; filename="shoppinglist.txt"'
+        response = HttpResponse(shopping_itog, content_type="text/plain")
+        response["Content-Disposition"] = 'attachment; filename="shoppinglist.txt"'
         return response
 
 
@@ -93,6 +85,7 @@ class RecipeFavoritesViewSet(viewsets.ModelViewSet):
     Добавление /
     удаление из списка
     """
+
     serializer_class = RecipeFavoritesSerializer
     queryset = RecipeFavorites.objects.all()
     permission_classes = [IsAuthenticated]
@@ -101,25 +94,23 @@ class RecipeFavoritesViewSet(viewsets.ModelViewSet):
         """Метод для добавления рецепта
         в список избранного
         """
-        recipes_id = self.kwargs['id']
+        recipes_id = self.kwargs["id"]
         recipes = get_object_or_404(Recipe, id=recipes_id)
-        RecipeFavorites.objects.create(
-            user=request.user,
-            recipes=recipes
-        )
+        RecipeFavorites.objects.create(user=request.user, recipes=recipes)
         serializer = RecipeFavoritesSerializer()
-        return Response(serializer.to_representation(instance=recipes),
-                        status=status.HTTP_201_CREATED)
+        return Response(
+            serializer.to_representation(instance=recipes),
+            status=status.HTTP_201_CREATED,
+        )
 
     def delete(self, request, *args, **kwargs):
         """Метод для удаления рецепта
         из списка избранного
         """
-        recipes_id = self.kwargs['id']
+        recipes_id = self.kwargs["id"]
         user_id = request.user.id
         RecipeFavorites.objects.filter(
-            user__id=user_id,
-            recipes__id=recipes_id
+            user__id=user_id, recipes__id=recipes_id
         ).delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -129,28 +120,24 @@ class FollowViewSet(viewsets.ModelViewSet):
     Cоздание подписки /
     удаление подписки
     """
+
     serializer_class = FollowSerializer
     pagination_class = CustomPagination
     permission_classes = [IsAuthenticated]
 
     def create(self, request, *args, **kwargs):
         """Метод для создания подписки"""
-        user_id = self.kwargs['id']
+        user_id = self.kwargs["id"]
         user = get_object_or_404(CustomUser, id=user_id)
-        subscribe = Follow.objects.create(
-            user=request.user, author=user)
-        serializer = FollowSerializer(subscribe,
-                                      context={'request': request})
+        subscribe = Follow.objects.create(user=request.user, author=user)
+        serializer = FollowSerializer(subscribe, context={"request": request})
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     def delete(self, request, *args, **kwargs):
         """Метод для удаления подписки"""
-        author_id = self.kwargs['id']
+        author_id = self.kwargs["id"]
         user_id = request.user.id
-        Follow.objects.filter(
-            user__id=user_id,
-            author__id=author_id
-        ).delete()
+        Follow.objects.filter(user__id=user_id, author__id=author_id).delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
@@ -160,6 +147,7 @@ class ShoppingViewSet(viewsets.ModelViewSet):
     удаление рецепта из списка покупок /
     скачивание списка покупок
     """
+
     serializer_class = ShoppingListSerializer
     pagination_class = CustomPagination
     queryset = ShoppingList.objects.all()
@@ -169,24 +157,20 @@ class ShoppingViewSet(viewsets.ModelViewSet):
         """Метод для добавление рецепта в
         список покупок
         """
-        recipe_id = self.kwargs['id']
+        recipe_id = self.kwargs["id"]
         recipes = get_object_or_404(Recipe, id=recipe_id)
-        ShoppingList.objects.create(
-            user=request.user,
-            recipes=recipes
-        )
+        ShoppingList.objects.create(user=request.user, recipes=recipes)
         serializer = ShoppingListSerializer()
-        return Response(serializer.to_representation(instance=recipes),
-                        status=status.HTTP_201_CREATED)
+        return Response(
+            serializer.to_representation(instance=recipes),
+            status=status.HTTP_201_CREATED,
+        )
 
     def delete(self, request, *args, **kwargs):
         """Метод для удаление рецепта из
         списка покупок
         """
-        recipe_id = self.kwargs['id']
+        recipe_id = self.kwargs["id"]
         user_id = request.user.id
-        ShoppingList.objects.filter(
-            user__id=user_id,
-            recipes__id=recipe_id
-        ).delete()
+        ShoppingList.objects.filter(user__id=user_id, recipes__id=recipe_id).delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
